Fixed LSTInputProducer
