+++ RecoTracker/LST/plugins/alpaka/LSTProducer.cc
@@ -66,7 +66,7 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE {
     }
 
   private:
-    device::EDGetToken<lst::LSTInputCollection> lstInputToken_;
+    device::EDGetToken<lst::LSTInputDeviceCollection> lstInputToken_;
     device::ESGetToken<lst::LSTESData<Device>, TrackerRecoGeometryRecord> lstESToken_;
     const bool verbose_;
     const double ptCut_;
