+++ RecoTracker/LSTCore/interface/alpaka/LSTInputDeviceCollection.h
@@ -1,0 +1,13 @@
+#ifndef RecoTracker_LSTCore_interface_alpaka_LSTInputDeviceCollection_h
+#define RecoTracker_LSTCore_interface_alpaka_LSTInputDeviceCollection_h
+
+#include "DataFormats/Portable/interface/alpaka/PortableCollection.h"
+
+#include "RecoTracker/LSTCore/interface/alpaka/Common.h"
+#include "RecoTracker/LSTCore/interface/LSTInputSoA.h"
+
+namespace ALPAKA_ACCELERATOR_NAMESPACE::lst {
+  using LSTInputDeviceCollection = PortableCollection3<InputHitsSoA, InputPixelHitsSoA, InputPixelSeedsSoA>;
+}  // namespace ALPAKA_ACCELERATOR_NAMESPACE::lst
+
+#endif
