+++ RecoTracker/LSTCore/src/alpaka/LSTEvent.dev.cc
@@ -78,7 +78,7 @@ void LSTEvent::resetEventSync() {
   modulesHC_.reset();
 }
 
-void LSTEvent::addInputToEvent(LSTInputCollection const* lstInputDC) {
+void LSTEvent::addInputToEvent(LSTInputDeviceCollection const* lstInputDC) {
   lstInputDC_ = lstInputDC;
 
   pixelSize_ = lstInputDC_->sizes()[2];
