+++ RecoTracker/LSTCore/src/alpaka/classes_cuda.h
@@ -3,6 +3,6 @@
 
 #include "DataFormats/Common/interface/DeviceProduct.h"
 #include "DataFormats/Common/interface/Wrapper.h"
-#include "RecoTracker/LSTCore/interface/alpaka/LSTInputCollection.h"
+#include "RecoTracker/LSTCore/interface/alpaka/LSTInputDeviceCollection.h"
 
 #endif
