+++ RecoTracker/LSTCore/src/alpaka/classes_rocm_def.xml
@@ -1,5 +1,5 @@
 <lcgdict>
-  <class name="alpaka_rocm_async::lst::LSTInputCollection" persistent="false"/>
-  <class name="edm::DeviceProduct<alpaka_rocm_async::lst::LSTInputCollection>" persistent="false"/>
-  <class name="edm::Wrapper<edm::DeviceProduct<alpaka_rocm_async::lst::LSTInputCollection>>" persistent="false"/>
+  <class name="alpaka_rocm_async::lst::LSTInputDeviceCollection" persistent="false"/>
+  <class name="edm::DeviceProduct<alpaka_rocm_async::lst::LSTInputDeviceCollection>" persistent="false"/>
+  <class name="edm::Wrapper<edm::DeviceProduct<alpaka_rocm_async::lst::LSTInputDeviceCollection>>" persistent="false"/>
 </lcgdict>
