+++ RecoTracker/LSTCore/standalone/code/core/trkCore.h
@@ -13,7 +13,7 @@
 #include <filesystem>
 
 using LSTEvent = ALPAKA_ACCELERATOR_NAMESPACE::lst::LSTEvent;
-using LSTInputDeviceCol = ALPAKA_ACCELERATOR_NAMESPACE::lst::LSTInputCollection;
+using LSTInputDeviceCollection = ALPAKA_ACCELERATOR_NAMESPACE::lst::LSTInputDeviceCollection;
 using ::lst::PixelType;
 
 // --------------------- ======================== ---------------------
@@ -53,7 +53,7 @@ TVector3 calculateR3FromPCA(const TVector3 &p3, const float dxy, const float dz)
 
 float addInputsToEventPreLoad(LSTEvent *event,
                               lst::LSTInputHostCollection *lstInputHC,
-                              LSTInputDeviceCol *lstInputDC,
+                              LSTInputDeviceCollection *lstInputDC,
                               ALPAKA_ACCELERATOR_NAMESPACE::Queue &queue);
 
 void printTimingInformation(std::vector<std::vector<float>> &timing_information, float fullTime, float fullavg);
